Module leo_asg::expression

This module defines an expression node in an asg.

Notable differences after conversion from an ast expression include:

Storing variable references instead of variable identifiers - better history tracking and mutability

Resolving constant values - optimizes execution of program circuit.

Modules

array_access 🔒

array_init 🔒

array_inline 🔒

array_range_access 🔒

binary 🔒

call 🔒

cast 🔒

circuit_access 🔒

circuit_init 🔒

constant 🔒

lengthof 🔒

ternary 🔒

tuple_access 🔒

tuple_init 🔒

unary 🔒

variable_ref 🔒

Structs

ArrayAccessExpression

ArrayInitExpression

ArrayInlineExpression

ArrayRangeAccessExpression

BinaryExpression

CallExpression

CastExpression

CircuitAccessExpression

CircuitInitExpression

Constant

LengthOfExpression

TernaryExpression

TupleAccessExpression

TupleInitExpression

UnaryExpression

VariableRef

Enums

BinaryOperation

BinaryOperationClass

Expression

UnaryOperation

Traits

AstNode

ExpressionNode