Module leo_asg::program

This module defines the program node for an asg.

Modules

alias 🔒

circuit 🔒

function 🔒

Structs

Alias

Annotation

Circuit

Function

Program

Stores the Leo program abstract semantic graph (ASG).

Enums

CircuitMember

FunctionQualifier

ImportSymbol 🔒

Enumerates what names are imported from a package.

Functions

check_top_level_namespaces 🔒

Checks whether a given string is found in any other global namespaces. If it is found it returns an error.

resolve_import_package 🔒

resolve_import_package_access 🔒