Module leo_asg::reducer

This module contains the reducer which iterates through ast nodes - converting them into asg nodes and saving relevant information.

Modules

monoid 🔒

monoidal_director 🔒

monoidal_reducer 🔒

reconstructing_director 🔒

reconstructing_reducer 🔒

visitor 🔒

visitor_director 🔒

Structs

BoolAnd

MonoidalDirector

ReconstructingDirector

SetAppend

VecAppend

VisitorDirector

Enums

VisitResult

Traits

ExpressionVisitor

Monoid

MonoidalReducerExpression

MonoidalReducerProgram

MonoidalReducerStatement

ProgramVisitor

ReconstructingReducerExpression

ReconstructingReducerProgram

ReconstructingReducerStatement

StatementVisitor

Type Definitions

ConcreteVisitResult