Struct leo_asg::scope::Scope

pub struct Scope<'a> {
    pub context: AsgContext<'a>,
    pub id: u32,
    pub parent_scope: Cell<Option<&'a Scope<'a>>>,
    pub function: Cell<Option<&'a Function<'a>>>,
    pub variables: RefCell<IndexMap<String, &'a Variable<'a>>>,
    pub functions: RefCell<IndexMap<String, &'a Function<'a>>>,
    pub global_consts: RefCell<IndexMap<String, &'a DefinitionStatement<'a>>>,
    pub circuits: RefCell<IndexMap<String, &'a Circuit<'a>>>,
    pub input: Cell<Option<Input<'a>>>,
}

An abstract data type that track the current bindings for variables, functions, and circuits.

Fields
context: AsgContext<'a>id: u32

The unique id of the scope.

parent_scope: Cell<Option<&'a Scope<'a>>>

The parent scope that this scope inherits.

function: Cell<Option<&'a Function<'a>>>

The function definition that this scope occurs in.

variables: RefCell<IndexMap<String, &'a Variable<'a>>>

Maps variable name => variable.

functions: RefCell<IndexMap<String, &'a Function<'a>>>

Maps function name => function.

global_consts: RefCell<IndexMap<String, &'a DefinitionStatement<'a>>>

Maps global constant name => global const code block.

circuits: RefCell<IndexMap<String, &'a Circuit<'a>>>

Maps circuit name => circuit.

input: Cell<Option<Input<'a>>>

The main input to the program.

Implementations

impl<'a> Scope<'a>

pub fn resolve_variable(&self, name: &str) -> Option<&'a Variable<'a>>

Returns a reference to the variable corresponding to the name.

If the current scope did not have this name present, then the parent scope is checked. If there is no parent scope, then None is returned.

pub fn resolve_current_function(&self) -> Option<&'a Function<'_>>

Returns a reference to the current function.

If the current scope did not have a function present, then the parent scope is checked. If there is no parent scope, then None is returned.

pub fn resolve_input(&self) -> Option<Input<'a>>

Returns a reference to the current input.

If the current scope did not have an input present, then the parent scope is checked. If there is no parent scope, then None is returned.

pub fn resolve_function(&self, name: &str) -> Option<&'a Function<'a>>

Returns a reference to the function corresponding to the name.

If the current scope did not have this name present, then the parent scope is checked. If there is no parent scope, then None is returned.

pub fn resolve_circuit(&self, name: &str) -> Option<&'a Circuit<'a>>

Returns a reference to the circuit corresponding to the name.

If the current scope did not have this name present, then the parent scope is checked. If there is no parent scope, then None is returned.

pub fn make_subscope(self: &'a Scope<'a>) -> &'a Scope<'a>

Returns a new scope given a parent scope.

pub fn resolve_ast_type(&self, type_: &Type, span: &Span) -> Result<Type<'a>>

Returns the type returned by the current scope.