Module leo_asg::statement

This module defines a statement node in an asg.

Ast statement nodes can be directly converted into asg nodes with no major differences.

Modules

assign

block

conditional

console

definition

expression

iteration

return_

Structs

AssignStatement

BlockStatement

ConditionalStatement

ConsoleArgs

ConsoleStatement

DefinitionStatement

ExpressionStatement

IterationStatement

ReturnStatement

Enums

AssignAccess

AssignOperation

ConsoleFunction

Statement